pyproject.toml
[tool.poetry]
name = "GenshinWikiMap"
version = "0.1.0"
description = "原神图鉴制作"
authors = ["CMHopeSunshine <277073121@qq.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.3.0"
pydantic = "^1.10.2"
ujson = "^5.5.0"
httpx = "^0.23.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
